+++ src/index.ts
@@ -1,2 +1,8 @@
-export * from './lib/index';
+export {
+  PerfectScrollbarComponent, PerfectScrollbarDirective,
+  PerfectScrollbarConfig, PerfectScrollbarConfigInterface,
+  provideDefaultConfig, provideForRootGuard,
+  PERFECT_SCROLLBAR_GUARD, PERFECT_SCROLLBAR_CONFIG,
+  PerfectScrollbarModule
+} from './lib/index';