+++ src/lib/index.ts
@@ -1,4 +1,13 @@
-export * from './perfect-scrollbar.component';
-export * from './perfect-scrollbar.directive';
-export * from './perfect-scrollbar.interfaces';
-export * from './perfect-scrollbar.module';
+export { PerfectScrollbarComponent } from './perfect-scrollbar.component';
+
+export { PerfectScrollbarDirective } from './perfect-scrollbar.directive';
+
+export {
+  PerfectScrollbarConfig, PerfectScrollbarConfigInterface
+} from './perfect-scrollbar.interfaces';
+
+export {
+  provideDefaultConfig, provideForRootGuard,
+  PERFECT_SCROLLBAR_GUARD, PERFECT_SCROLLBAR_CONFIG,
+  PerfectScrollbarModule
+} from './perfect-scrollbar.module';