+++ src/lib/perfect-scrollbar.directive.ts
@@ -1,6 +1,6 @@
 import * as Ps from 'perfect-scrollbar';
 
-import { Directive, DoCheck, OnDestroy, Input, Optional, ElementRef, AfterViewInit, NgZone } from '@angular/core';
+import { Directive, DoCheck, OnDestroy, OnChanges, Input, Optional, HostBinding, ElementRef, AfterViewInit, SimpleChanges, NgZone } from '@angular/core';
 
 import { PerfectScrollbarConfig, PerfectScrollbarConfigInterface } from './perfect-scrollbar.interfaces';
 
@@ -10,12 +10,15 @@ import { PerfectScrollbarConfig, PerfectScrollbarConfigInterface } from './perfe
     style: 'position: relative;'
   }
 })
-export class PerfectScrollbarDirective implements DoCheck, OnDestroy, AfterViewInit {
+export class PerfectScrollbarDirective implements DoCheck, OnDestroy, OnChanges, AfterViewInit {
   private width: number;
   private height: number;
 
   private contentWidth: number;
   private contentHeight: number;
+
+  @HostBinding('hidden')
+  @Input() hidden: boolean = false;
 
   @Input() runInsideAngular: boolean = false;
 
@@ -55,6 +58,12 @@ export class PerfectScrollbarDirective implements DoCheck, OnDestroy, AfterViewI
       this.zone.runOutsideAngular(() => {
         Ps.destroy(this.elementRef.nativeElement);
       });
+    }
+  }
+
+  ngOnChanges(changes: SimpleChanges) {
+    if (changes['hidden'] && !this.hidden) {
+      this.update();
     }
   }