+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+2.5.0 Last 2.x release
+
+  - Preparing for moving to Angular 4.x
+
+2.0.8 Name change release
+
+  - Changed from angular2 to ngx prefix
+
 2.0.6 Bug fix release
 
   - Added better support for flex-layout